+++ delphi-frontend/src/main/java/au/com/integradev/delphi/DelphiProperties.java
@@ -20,6 +20,7 @@ package au.com.integradev.delphi;
 
 import au.com.integradev.delphi.compiler.CompilerVersion;
 import au.com.integradev.delphi.compiler.Toolchain;
+import au.com.integradev.delphi.core.Delphi;
 import java.util.List;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
@@ -28,24 +29,30 @@ import org.sonar.api.config.PropertyDefinition;
 import org.sonar.api.resources.Qualifiers;
 
 public final class DelphiProperties {
-  public static final Toolchain COMPILER_TOOLCHAIN_DEFAULT = Toolchain.DCC32;
-  public static final CompilerVersion COMPILER_VERSION_DEFAULT =
-      CompilerVersion.fromVersionSymbol("VER340");
-
-  public static final String SEARCH_PATH_KEY = "sonar.delphi.sources.searchPath";
-  public static final String BDS_PATH_KEY = "sonar.delphi.bds.path";
-  public static final String COMPILER_TOOLCHAIN_KEY = "sonar.delphi.compiler.toolchain";
-  public static final String COMPILER_VERSION_KEY = "sonar.delphi.compiler.version";
+  public static final String BDS_PATH_KEY = "sonar.delphi.bdsPath";
+  public static final String COMPILER_TOOLCHAIN_KEY = "sonar.delphi.toolchain";
+  public static final String COMPILER_VERSION_KEY = "sonar.delphi.compilerVersion";
+  public static final String SEARCH_PATH_KEY = "sonar.delphi.searchPath";
   public static final String CONDITIONAL_DEFINES_KEY = "sonar.delphi.conditionalDefines";
   public static final String CONDITIONAL_UNDEFINES_KEY = "sonar.delphi.conditionalUndefines";
   public static final String UNIT_SCOPE_NAMES_KEY = "sonar.delphi.unitScopeNames";
   public static final String UNIT_ALIASES_KEY = "sonar.delphi.unitAliases";
-  public static final String COVERAGE_TOOL_KEY = "sonar.delphi.coverage.tool";
-  public static final String COVERAGE_REPORT_KEY = "sonar.delphi.coverage.reportPaths";
+  public static final String TEST_TYPE_KEY = "sonar.delphi.testType";
   public static final String NUNIT_REPORT_PATHS_PROPERTY = "sonar.delphi.nunit.reportPaths";
-  public static final String TEST_SUITE_TYPE_KEY = "sonar.delphi.testSuiteType";
+  public static final String COVERAGE_REPORT_KEY = "sonar.delphi.coverage.reportPaths";
 
-  public static final String COVERAGE_TOOL_DELPHI_CODE_COVERAGE = "dcc";
+  private static final String DELPHI_CATEGORY = "delphi";
+  private static final String GENERAL_SUBCATEGORY = "General";
+  private static final String TOOLCHAIN_SUBCATEGORY = "Toolchain";
+  private static final String COMPILER_OPTIONS_SUBCATEGORY = "Compiler Options";
+  private static final String TEST_SUBCATEGORY = "Test and Coverage";
+
+  private static final String BDS_PATH_DEFAULT =
+      "C:\\Program Files (x86)\\Embarcadero\\Studio\\22.0";
+  public static final Toolchain COMPILER_TOOLCHAIN_DEFAULT = Toolchain.DCC32;
+  public static final CompilerVersion COMPILER_VERSION_DEFAULT =
+      CompilerVersion.fromVersionSymbol("VER350");
+  private static final String TEST_TYPE_DEFAULT = "TestFramework.TTestCase";
 
   private DelphiProperties() {
     // hide public constructor
@@ -53,96 +60,124 @@ public final class DelphiProperties {
 
   public static List<PropertyDefinition> getProperties() {
     return List.of(
-        PropertyDefinition.builder(DelphiProperties.SEARCH_PATH_KEY)
-            .name("Search path")
-            .description("Directories to search in for include files and unit imports.")
+        PropertyDefinition.builder(Delphi.FILE_SUFFIXES_KEY)
+            .category(DELPHI_CATEGORY)
+            .subCategory(GENERAL_SUBCATEGORY)
+            .defaultValue(Delphi.DEFAULT_FILE_SUFFIXES)
+            .name("File suffixes")
+            .description(
+                "List of suffixes for Delphi files to analyze."
+                    + " To not filter, leave the list empty.")
             .multiValues(true)
             .onQualifiers(Qualifiers.PROJECT)
             .build(),
         PropertyDefinition.builder(DelphiProperties.BDS_PATH_KEY)
+            .category(DELPHI_CATEGORY)
+            .subCategory(TOOLCHAIN_SUBCATEGORY)
+            .defaultValue(BDS_PATH_DEFAULT)
             .name("BDS path")
-            .description(
-                "Path to the Delphi BDS folder."
-                    + "Example: C:\\Program Files (x86)\\Embarcadero\\Studio\\20.0")
+            .description("Path to the Delphi BDS folder.")
             .onQualifiers(Qualifiers.PROJECT)
             .build(),
         PropertyDefinition.builder(DelphiProperties.COMPILER_TOOLCHAIN_KEY)
-            .name("Compiler toolchain")
+            .category(DELPHI_CATEGORY)
+            .subCategory(TOOLCHAIN_SUBCATEGORY)
             .defaultValue(COMPILER_TOOLCHAIN_DEFAULT.name())
+            .name("Compiler toolchain")
             .description(
-                "The compiler toolchain used by this project. Options are: "
+                "The compiler toolchain. Options: "
                     + StringUtils.join(
                         Stream.of(Toolchain.values())
-                            .map(value -> "\"" + value + "\"")
+                            .map(value -> "`" + value + "`")
                             .collect(Collectors.toList())))
             .onQualifiers(Qualifiers.PROJECT)
             .build(),
         PropertyDefinition.builder(DelphiProperties.COMPILER_VERSION_KEY)
-            .name("Compiler version")
+            .category(DELPHI_CATEGORY)
+            .subCategory(TOOLCHAIN_SUBCATEGORY)
             .defaultValue(COMPILER_VERSION_DEFAULT.symbol())
+            .name("Compiler version")
             .description(
                 "The Delphi conditional symbol representing the compiler version."
-                    + " Format is \"VER&lt;nnn&gt;\"."
-                    + " See: http://docwiki.embarcadero.com/RADStudio/en/Compiler_Versions")
+                    + " Format: `VER&lt;nnn&gt;`.")
+            .onQualifiers(Qualifiers.PROJECT)
+            .build(),
+        PropertyDefinition.builder(DelphiProperties.SEARCH_PATH_KEY)
+            .category(DELPHI_CATEGORY)
+            .subCategory(COMPILER_OPTIONS_SUBCATEGORY)
+            .name("Search path")
+            .description(
+                "List of directories to search for include files and unit imports."
+                    + " Each path may be absolute or relative to the project base directory.")
+            .multiValues(true)
             .onQualifiers(Qualifiers.PROJECT)
             .build(),
         PropertyDefinition.builder(DelphiProperties.CONDITIONAL_DEFINES_KEY)
-            .name("Conditional Defines")
-            .description("List of conditional defines to define while parsing the project")
+            .category(DELPHI_CATEGORY)
+            .subCategory(COMPILER_OPTIONS_SUBCATEGORY)
+            .name("Conditional defines")
+            .description(
+                "List of conditional defines to define while parsing the project, in addition to"
+                    + " the defines aggregated from the project files.")
             .multiValues(true)
             .onQualifiers(Qualifiers.PROJECT)
             .build(),
         PropertyDefinition.builder(DelphiProperties.CONDITIONAL_UNDEFINES_KEY)
-            .name("Conditional Undefines")
+            .category(DELPHI_CATEGORY)
+            .subCategory(COMPILER_OPTIONS_SUBCATEGORY)
+            .name("Conditional undefines")
             .description(
-                "List of conditional defines to undefine before parsing the project. This is useful"
-                    + " if you aggregate the defines from your project files, but still want to"
-                    + " exclude certain ones.")
+                "List of conditional defines to consider undefined while parsing the project."
+                    + " This is useful for flicking off some specific defines that were aggregated"
+                    + " from the project files.")
             .multiValues(true)
             .onQualifiers(Qualifiers.PROJECT)
             .build(),
         PropertyDefinition.builder(DelphiProperties.UNIT_SCOPE_NAMES_KEY)
-            .name("Unit Scope Names")
-            .description("List of Unit scope names, used for unit import resolution.")
+            .category(DELPHI_CATEGORY)
+            .subCategory(COMPILER_OPTIONS_SUBCATEGORY)
+            .name("Unit scope names")
+            .description("List of unit scope names, used for import resolution.")
             .multiValues(true)
             .onQualifiers(Qualifiers.PROJECT)
             .build(),
         PropertyDefinition.builder(DelphiProperties.UNIT_ALIASES_KEY)
-            .name("Unit Aliases")
+            .category(DELPHI_CATEGORY)
+            .subCategory(COMPILER_OPTIONS_SUBCATEGORY)
+            .name("Unit aliases")
             .description(
-                "List of Unit Aliases, used for unit import resolution."
-                    + " NOTE: Each Unit Alias should follow this format: 'AliasName=UnitName'")
+                "List of unit aliases, used for import resolution. Format: `AliasName=UnitName`")
             .multiValues(true)
             .onQualifiers(Qualifiers.PROJECT)
             .build(),
-        PropertyDefinition.builder(DelphiProperties.COVERAGE_TOOL_KEY)
-            .name("Coverage Tool")
-            .defaultValue(DelphiProperties.COVERAGE_TOOL_DELPHI_CODE_COVERAGE)
+        PropertyDefinition.builder(DelphiProperties.TEST_TYPE_KEY)
+            .category(DELPHI_CATEGORY)
+            .subCategory(TEST_SUBCATEGORY)
+            .defaultValue(TEST_TYPE_DEFAULT)
+            .name("Test Type")
             .description(
-                "Used coverage tool. Options are: \""
-                    + DelphiProperties.COVERAGE_TOOL_DELPHI_CODE_COVERAGE
-                    + "\"")
-            .onQualifiers(Qualifiers.PROJECT)
-            .build(),
-        PropertyDefinition.builder(DelphiProperties.COVERAGE_REPORT_KEY)
-            .name("Coverage Report Paths")
-            .description("List of paths containing coverage reports from the specified tool")
-            .multiValues(true)
+                "A fully qualified type name. Any code within this type or its descendants will be"
+                    + " treated as test code.")
             .onQualifiers(Qualifiers.PROJECT)
             .build(),
         PropertyDefinition.builder(DelphiProperties.NUNIT_REPORT_PATHS_PROPERTY)
-            .name("NUnit Report Paths")
-            .description("List of paths to NUnit report directories")
+            .category(DELPHI_CATEGORY)
+            .subCategory(TEST_SUBCATEGORY)
+            .name("Path to NUnit report(s)")
+            .description(
+                "List of directories containing the *.xml NUnit report files."
+                    + " Each path may be absolute or relative to the project base directory.")
             .multiValues(true)
             .onQualifiers(Qualifiers.PROJECT)
             .build(),
-        PropertyDefinition.builder(DelphiProperties.TEST_SUITE_TYPE_KEY)
-            .name("Test Suite Type")
-            .defaultValue("")
+        PropertyDefinition.builder(DelphiProperties.COVERAGE_REPORT_KEY)
+            .category(DELPHI_CATEGORY)
+            .subCategory(TEST_SUBCATEGORY)
+            .name("Path to coverage report(s)")
             .description(
-                "Rules can be configured not to apply to test code. Any type that inherits from the"
-                    + " Test Suite type will have its methods considered as test code."
-                    + " NOTE: A fully qualified type name is expected.")
+                "List of directories containing the *.xml Delphi Code Coverage report files."
+                    + " Each path may be absolute or relative to the project base directory")
+            .multiValues(true)
             .onQualifiers(Qualifiers.PROJECT)
             .build());
   }